move all low gas functions to low gas script file
